refactor(observability): downshift server http_client_pool to observability (P4a Batch 1c)
Move aragora.server.http_client_pool DOWN to aragora.observability.http_client_pool
so foundation/infrastructure modules reach the shared HTTP client pool without
importing aragora.server. The old path stays a sys.modules-alias DeprecationWarning
shim (import-compatible; old-path monkeypatches resolve to the moved module).

Repoint the foundation/infra importers: observability/alerting.py, observability/slo.py,
billing/budget_alert_notifier.py, queue/webhook_worker.py. Domain/application callers
intentionally stay on the shim (by-design two-sided VAL-CROSS-004 signal).

Validation: make test-smoke exit 0 (Core/Server/Memory canary, VAL-P4A-013);
targeted http-pool + webhook tests 454 passed; new deprecation test green;
check_import_contracts --layers foundation,infrastructure adds no new violation
(only the ambient utils->caching). No baseline shrink: observability/billing/queue ->
server all retain indirect/partial residuals (LAST-CONTRIBUTOR rule).
